[refactor] #1 신고 코드 중 시설물 관련 코드 수정
- 새로 바뀐 Facility 객체를 사용하도록 수정함
